Move GeneratedObjectMerger to its own class
