+++ generator/src/main/java/com/scottlogic/deg/generator/generation/databags/GeneratedObjectMerger.java
@@ -1,0 +1,32 @@
+package com.scottlogic.deg.generator.generation.databags;
+
+import com.scottlogic.deg.generator.DataBagValue;
+import com.scottlogic.deg.generator.Field;
+import com.scottlogic.deg.generator.FlatMappingSpliterator;
+import com.scottlogic.deg.generator.ProfileFields;
+import com.scottlogic.deg.generator.outputs.CellSource;
+import com.scottlogic.deg.generator.outputs.RowSource;
+
+import java.util.*;
+import java.util.stream.Collectors;
+
+
+public class GeneratedObjectMerger {
+    public static final GeneratedObject empty = new GeneratedObject(new HashMap<>());
+
+    public static GeneratedObject merge(GeneratedObject... bags) {
+        Map<Field, DataBagValue> newFieldToValue = new HashMap<>();
+
+        FlatMappingSpliterator.flatMap(Arrays.stream(bags)
+            .map(r -> r.getFieldToValue().entrySet().stream()),
+            entrySetStream -> entrySetStream)
+            .forEach(entry -> {
+                if (newFieldToValue.containsKey(entry.getKey()))
+                    throw new IllegalArgumentException("Databags can't be merged because they overlap on field " + entry.getKey().name);
+
+                newFieldToValue.put(entry.getKey(), entry.getValue());
+            });
+
+        return new GeneratedObject(newFieldToValue);
+    }
+}
